+++ flag/service/mimir/mimir.go
@@ -1,0 +1,5 @@
+package mimir
+
+type Mimir struct {
+	Enabled string
+}