+++ flag/service/service.go
@@ -6,6 +6,7 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/v2/flag/service/grafana"
 	"github.com/giantswarm/prometheus-meta-operator/v2/flag/service/ingress"
 	"github.com/giantswarm/prometheus-meta-operator/v2/flag/service/installation"
+	"github.com/giantswarm/prometheus-meta-operator/v2/flag/service/mimir"
 	"github.com/giantswarm/prometheus-meta-operator/v2/flag/service/opsgenie"
 	"github.com/giantswarm/prometheus-meta-operator/v2/flag/service/prometheus"
 	"github.com/giantswarm/prometheus-meta-operator/v2/flag/service/provider"
@@ -21,6 +22,7 @@ type Service struct {
 	Installation installation.Installation
 	Kubernetes   kubernetes.Kubernetes
 	Opsgenie     opsgenie.Opsgenie
+	Mimir        mimir.Mimir
 	Prometheus   prometheus.Prometheus
 	Provider     provider.Provider
 	Security     security.Security