+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Disable rule evaluation in Prometheus if mimir is enabled.
+
 ## [4.68.3] - 2024-02-19
 
 ### Fixed